Add join game and basic Lobby to UI
Some server-side fixes, including switching the date over to a Long so
we can use Posix values directly in the client.

Added some tests for decoding server messages. The messages are
hard-coded so I need to remember to update them, be good to find a way
to automate populating the tests with up-to-date meesage JSON strings?

The sharedDb setting in build.sbt lets us use the Local DynamoDB web
shell to inspect the devserver's database, which is *super useful*.