Refactor winnings datastructures and implement first draft
The winnings are now calculated and described "by pot" instead of by
player. This gives much more helpful information and is easier to
calculate. It's also very straightforward to go from "by pot" to "by
player" since the latter is a subset, information-wise.

This led to an overhaul of the Client-facing datastructures for
conveying showdown results, which is done. Some more work will follow
actually fitting this into the public endpoint.

An implementation of the winnings logic exists, and a couple of very
basic tests verify that this is a commitable first approximation. The
edge cases will reveal bugs so more tests will follow!