Add logic for resetting blinds and dealer between rounds
Tricky logic for working out where the blinds are for the next round,
as well as which player should be dealer.

We use the "big blind moves" rules for determining blinds and dealers,
with the additional rule that busted players are never the dealer (the
latter does not affect gameplay). This means the Big Blind always
moves correctly round the table and the small blind and dealer will
move around that to make sure that players:
* do not skip blinds
* are never required to repeat a blind

This can result in rounds where there is no small blind, as well as
situations where the dealer stays, or even moves backwards. The focus
is on making sure that blinds move fairly.

There are several other fixes and lots of tests to accompany this
change.