Fix a number of gameplay bugs (and introduces some new ones)
Most notably, add the distinction between a new round and subsequent
phases of the round, for turn order. If blinds have been dealt there's
a different setup to the turn order after cards are dealt. This
distinction was missing.

Some other bugfixes, but there are still some errors to sort out.