Tidy up event/errors overlay code
The overlay UI logic now makes a bit of sense instead of being
scattered around.

The events element is moved out of the way of important UI
functionality.

Also, when both events and errors are showing, touchevents are removed
from the fixed container and re-added to each content element. This
prevents the fixed inFront element from hogging all page interactions.