First draft of timer support in the frontend
This adds a visual representation of the current and upcoming timer
status to the UI, unstyled for now.

Also improves the edit timer UI element in the lobby.

There'sa fiddly function for working out how far through the timer
leves we are, using the current time. This function has some tests.