Add the ability to edit blinds and timers
Introduces the UI concept of "overlays" to manage this more
easily. This allows blinds to be edited from any (relevant) point in
the application lifecycle, without having to wire it in everywhere.

The overlay concept also allows for a help text overlay, which has
been populated with some placehodler content for now.

Keyboard shortcuts have been added to the app, to allow control of the
overlay. Additional keyboard functionality should follow in the
future, for controlling the game itself on devices with keyboards.

The update blinds API call has been updated to add support for editing
blind timers, including the ability to jump to specific points in a
running timer.

Also fixes a bug that I was seeing locally where the app had not
properly disconnected from the API websocket on reload, preventing
reconnection. Using a page unload handler to disconnect appears to
resolve the problem.