Always send messages to the active address for the active player
The current player is always on the "current" connection, with an
active and correct address. Th eplayer address in the database may be
out of date, in contrast.

This forces us to always use the active address when sending messages
to the player that performed the current action. This dramatically
reduces the chances of a message not being properly delivered to the
active player.