First draft of game event logs
These logs should provide a complete picture of how games
played out, which is useful for analysis, AI development etc.

They also allow the server to know what just happened, which is
crucial for supporting "undo".

Since the data in the log is (currently) persisted indefinitely, it is
pseudonymous until the game data itself expires at which point the
data is then anonymous.